+++ app/src/main/java/com/decoutkhanqindev/dexreader/domain/usecase/settings/ObserveContentLanguageUseCase.kt
@@ -1,0 +1,14 @@
+package com.decoutkhanqindev.dexreader.domain.usecase.settings
+
+import com.decoutkhanqindev.dexreader.domain.entity.value.manga.MangaLanguage
+import com.decoutkhanqindev.dexreader.domain.repository.settings.SettingsRepository
+import com.decoutkhanqindev.dexreader.util.CoroutineHandler.toFlowResult
+import kotlinx.coroutines.flow.Flow
+import javax.inject.Inject
+
+class ObserveContentLanguageUseCase @Inject constructor(
+  private val repository: SettingsRepository,
+) {
+  operator fun invoke(): Flow<Result<MangaLanguage>> =
+    repository.observeContentLanguage().toFlowResult()
+}