+++ app/src/main/java/com/decoutkhanqindev/dexreader/domain/usecase/settings/SaveContentLanguageUseCase.kt
@@ -1,0 +1,13 @@
+package com.decoutkhanqindev.dexreader.domain.usecase.settings
+
+import com.decoutkhanqindev.dexreader.domain.entity.value.manga.MangaLanguage
+import com.decoutkhanqindev.dexreader.domain.repository.settings.SettingsRepository
+import com.decoutkhanqindev.dexreader.util.CoroutineHandler.runSuspendResultCatching
+import javax.inject.Inject
+
+class SaveContentLanguageUseCase @Inject constructor(
+  private val repository: SettingsRepository,
+) {
+  suspend operator fun invoke(value: MangaLanguage): Result<Unit> =
+    runSuspendResultCatching { repository.saveContentLanguage(value) }
+}